Booking & Payment
I haven't put any validation.
Confirmation done.
Events page.

Need to make content for events page.
and coming soon movies.

+++ CustomerMVC/Models/Booking.cs
@@ -20,7 +20,7 @@ namespace CustomerMVC.Models
         public string Movie_Time { get; set; }
         public string Session_Time { get; set; }
         public Nullable<decimal> Price { get; set; }
-        public Nullable<decimal> Ticket { get; set; }
+        public Nullable<int> Ticket { get; set; }
         public string First_Name { get; set; }
         public string Last_Name { get; set; }
         public string Email { get; set; }
